Move the code where we apply the correct TZ to the upper component
